requirements.txt
Flask==1.1.1
python-dotenv==0.10.3
dialogflow==0.7.2
pickledb
python-i18n
